requirements.txt
python-telegram-bot==20.7
Flask==3.0.3
python-dotenv==1.0.1
APScheduler==3.10.4